requirements.txt
streamlit>=1.28.0
langgraph>=0.0.25
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
python-dotenv>=1.0.0
python-docx>=1.0.0
fpdf>=1.7.2
